restructured CSS following the SEM and BIO methodology.
While I implemented this if everyone finds they hate it for some reason it can be changed, I just wanted to follow a structure if multiple people were going to be digging in.

If you don't feel you get this way of doing CSS, that's okay just write css how you want to and others will help correct the class names and whatnot, it's more important to make the extension do things than for the code to be pretty/readable.